pypi: rename fasterai to deoldify
It will be better to do:

    from deoldify import visualize

Than:

    from fasterai import visualize

The PyPI package will be called DeOldify, in this case
it makes more sense to have an import name that matches
the package name.

Also, fasterai resembles fastai library, and DeOldify
is not a library to use with fastai, it's built on
top of it.
